+++ ncplib/__init__.py
@@ -54,7 +54,7 @@ More information
 """
 
 
-__version__ = (2, 2, 1)
+__version__ = (2, 3, 0)
 
 
 from ncplib.app import *  # noqa